+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/flowcontroller/FlowControllerFactory.kt
@@ -30,15 +30,15 @@ internal class FlowControllerFactory(
         paymentOptionCallback: PaymentOptionCallback,
         paymentResultCallback: PaymentSheetResultCallback
     ) : this(
-        activity,
-        activity.lifecycleScope,
-        activity,
-        activity.applicationContext,
-        activity,
-        { activity.window.statusBarColor },
-        PaymentOptionFactory(activity.resources, StripeImageLoader(activity.applicationContext)),
-        paymentOptionCallback,
-        paymentResultCallback
+        viewModelStoreOwner = activity,
+        lifecycleScope = activity.lifecycleScope,
+        lifecycleOwner = activity,
+        appContext = activity.applicationContext,
+        activityResultCaller = activity,
+        statusBarColor = { activity.window.statusBarColor },
+        paymentOptionFactory = PaymentOptionFactory(activity.resources, StripeImageLoader(activity.applicationContext)),
+        paymentOptionCallback = paymentOptionCallback,
+        paymentResultCallback = paymentResultCallback,
     )
 
     constructor(
@@ -46,15 +46,15 @@ internal class FlowControllerFactory(
         paymentOptionCallback: PaymentOptionCallback,
         paymentResultCallback: PaymentSheetResultCallback
     ) : this(
-        fragment,
-        fragment.lifecycleScope,
-        fragment,
-        fragment.requireContext(),
-        fragment,
-        { fragment.activity?.window?.statusBarColor },
-        PaymentOptionFactory(fragment.resources, StripeImageLoader(fragment.requireContext().applicationContext)),
-        paymentOptionCallback,
-        paymentResultCallback
+        viewModelStoreOwner = fragment,
+        lifecycleScope = fragment.lifecycleScope,
+        lifecycleOwner = fragment,
+        appContext = fragment.requireContext().applicationContext,
+        activityResultCaller = fragment,
+        statusBarColor = { fragment.activity?.window?.statusBarColor },
+        paymentOptionFactory = PaymentOptionFactory(fragment.resources, StripeImageLoader(fragment.requireContext().applicationContext)),
+        paymentOptionCallback = paymentOptionCallback,
+        paymentResultCallback = paymentResultCallback,
     )
 
     fun create(): PaymentSheet.FlowController =